Restructure staking
