+++ Cargo.toml
@@ -6,6 +6,7 @@ members = [
 	"core/*",
 	"node",
 	"pallet/*",
+	"pallet/*/traits",
 	"precompile/*",
 	"runtime/*",
 ]
@@ -62,6 +63,7 @@ darwinia-precompile-staking       = { path = "precompile/staking", default-featu
 darwinia-precompile-state-storage = { path = "precompile/state-storage", default-features = false }
 darwinia-runtime                  = { path = "runtime/darwinia" }
 darwinia-staking                  = { path = "pallet/staking", default-features = false }
+darwinia-staking-traits           = { path = "pallet/staking/traits", default-features = false }
 dc-inflation                      = { path = "core/inflation", default-features = false }
 dc-primitives                     = { path = "core/primitives", default-features = false }
 dc-types                          = { path = "core/types" }
