+++ pallet/staking/traits/Cargo.toml
@@ -1,0 +1,28 @@
+[package]
+authors.workspace = true
+description       = "Darwinia parachain staking's traits."
+edition.workspace = true
+name              = "darwinia-staking-traits"
+readme            = "README.md"
+version.workspace = true
+
+[dependencies]
+# crates.io
+codec      = { workspace = true, package = "parity-scale-codec" }
+scale-info = { workspace = true }
+
+# subtrate
+sp-runtime = { workspace = true }
+
+[features]
+default = ["std"]
+std = [
+	# crates.io
+	"codec/std",
+	"scale-info/std",
+
+	# subtrate
+	"sp-runtime/std",
+]
+
+runtime-benchmarks = []
